// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const TemplateFileType = {
  PDF: 'PDF',
  PNG: 'PNG',
  JPG: 'JPG'
} as const

export type TemplateFileType = (typeof TemplateFileType)[keyof typeof TemplateFileType]


export const BatchStatus = {
  DRAFT: 'DRAFT',
  READY: 'READY',
  GENERATING: 'GENERATING',
  GENERATED: 'GENERATED',
  PUBLISHED: 'PUBLISHED',
  FAILED: 'FAILED'
} as const

export type BatchStatus = (typeof BatchStatus)[keyof typeof BatchStatus]


export const CertificateStatus = {
  PENDING: 'PENDING',
  GENERATING: 'GENERATING',
  GENERATED: 'GENERATED',
  FAILED: 'FAILED'
} as const

export type CertificateStatus = (typeof CertificateStatus)[keyof typeof CertificateStatus]